refactor(devs-infra): restructure project
