+++ .gitignore
@@ -1,8 +1,5 @@
 node_modules
-/build/
-InlineFilesTransformer.js
-StripStylesTransformer.js
-TransformUtils.js
+/src/build
 *.log
 .idea
 *.tgz
