--- .npmignore
@@ -1,32 +1,0 @@
-# Tests
-e2e
-
-# sources are inlined
-src
-
-# Developement scripts
-scripts
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# Dependency directories
-node_modules
-
-# Editor config
-.vscode
-.idea
-.editorconfig
-
-# CI
-.circleci
-
-# Others
-*.tgz
-.npmignore
-.npmrc
-greenkeeper.json
-jest.config.js
-tsconfig.json
-.github
-.gitattributes
