+++ e2e/test-app-v10/package.json
@@ -10,29 +10,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.3",
-    "@angular/common": "~10.0.3",
-    "@angular/compiler": "~10.0.3",
-    "@angular/core": "~10.0.3",
-    "@angular/forms": "~10.0.3",
-    "@angular/platform-browser": "~10.0.3",
-    "@angular/platform-browser-dynamic": "~10.0.3",
-    "@angular/router": "~10.0.3",
-    "rxjs": "~6.5.5",
-    "tslib": "^2.0.0",
+    "@angular/animations": "~10.0.8",
+    "@angular/common": "~10.0.8",
+    "@angular/compiler": "~10.0.8",
+    "@angular/core": "~10.0.8",
+    "@angular/forms": "~10.0.8",
+    "@angular/platform-browser": "~10.0.8",
+    "@angular/platform-browser-dynamic": "~10.0.8",
+    "@angular/router": "~10.0.8",
+    "rxjs": "~6.6.2",
+    "tslib": "^2.0.1",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.2",
-    "@angular/cli": "~10.0.2",
-    "@angular/compiler-cli": "~10.0.3",
-    "@types/jest": "^26.0.4",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "~0.1000.5",
+    "@angular/cli": "~10.0.5",
+    "@angular/compiler-cli": "~10.0.8",
+    "@types/jest": "^26.0.9",
+    "@types/node": "^12.12.54",
     "codelyzer": "^6.0.0",
-    "jest": "^26.1.0",
-    "jest-preset-angular": "file:../../",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "jest": "^26.3.0",
+    "jest-preset-angular": "file:../../src",
+    "ts-node": "~8.10.2",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7"
   }
 }
