+++ e2e/test-app-v9/package.json
@@ -14,28 +14,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.7",
-    "@angular/common": "~9.1.7",
-    "@angular/compiler": "~9.1.7",
-    "@angular/core": "~9.1.7",
-    "@angular/forms": "~9.1.7",
-    "@angular/platform-browser": "~9.1.7",
-    "@angular/platform-browser-dynamic": "~9.1.7",
-    "@angular/router": "~9.1.7",
-    "rxjs": "~6.5.5",
-    "tslib": "^1.10.0",
+    "@angular/animations": "~9.1.12",
+    "@angular/common": "~9.1.12",
+    "@angular/compiler": "~9.1.12",
+    "@angular/core": "~9.1.12",
+    "@angular/forms": "~9.1.12",
+    "@angular/platform-browser": "~9.1.12",
+    "@angular/platform-browser-dynamic": "~9.1.12",
+    "@angular/router": "~9.1.12",
+    "rxjs": "~6.6.2",
+    "tslib": "^1.13.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.6",
-    "@angular/cli": "~9.1.6",
-    "@angular/compiler-cli": "~9.1.7",
-    "@angular/language-service": "~9.1.7",
-    "@types/jest": "^26.0.4",
+    "@angular-devkit/build-angular": "~0.901.12",
+    "@angular/cli": "~9.1.12",
+    "@angular/compiler-cli": "~9.1.12",
+    "@angular/language-service": "~9.1.12",
+    "@types/jest": "^26.0.9",
     "@types/node": "^14.0.1",
-    "jest": "^26.0.1",
-    "jest-preset-angular": "file:../../",
-    "ts-node": "^8.10.1",
+    "jest": "^26.3.0",
+    "jest-preset-angular": "file:../../src",
+    "ts-node": "^8.10.2",
     "typescript": "~3.8.3"
   }
 }
