--- index.js
@@ -1,1 +1,0 @@
-module.exports = require('./build/setupJest.js');
