--- scripts/clean-build.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-const { removeSync } = require('fs-extra')
-const Paths = require('./lib/paths')
-
-removeSync(Paths.distDir)
