+++ src/.npmignore
@@ -1,0 +1,10 @@
+# Tests
+__tests__
+
+# sources are inlined
+zone-patch
+AngularNoNgAttributesSnapshotSerializer.js
+AngularSnapshotSerializer.js
+HTMLCommentSerializer.js
+*.ts
+*.tgz
