+++ src/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "jest-preset-angular",
+  "version": "8.2.1",
+  "description": "Jest preset configuration for Angular projects",
+  "main": "index.js",
+  "repository": "git@github.com:thymikee/jest-preset-angular.git",
+  "homepage": "https://github.com/thymikee/jest-preset-angular",
+  "author": "Michał Pierzchała <thymikee@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "pretty-format": "26.x",
+    "ts-jest": "26.x"
+  },
+  "peerDependencies": {
+    "@angular/core": ">=2.0.0",
+    "@angular/platform-browser-dynamic": ">=2.0.0",
+    "jest": "^26.0.0"
+  },
+  "engines": {
+    "node": ">= 10"
+  }
+}
