+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "es5",
     "baseUrl": ".",
     "module": "commonjs",
-    "outDir": "./build",
+    "outDir": "./src/build",
     "strict": true,
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
@@ -18,5 +18,5 @@
     "types": ["node"]
   },
   "include": ["src"],
-  "exclude": ["**/__tests__"]
+  "exclude": ["**/__tests__", "src/index.js", "src/jest-preset.js"]
 }
